Move group_hash to sapling module.
