+++ zcash_primitives/src/constants.rs
@@ -280,7 +280,7 @@ mod tests {
     use jubjub::SubgroupPoint;
 
     use super::*;
-    use crate::group_hash::group_hash;
+    use crate::sapling::group_hash::group_hash;
 
     fn find_group_hash(m: &[u8], personalization: &[u8; 8]) -> SubgroupPoint {
         let mut tag = m.to_vec();
